Added copyright headers, renamed `checkAndMutate()` to `checkAndUpdate()`.
