+++ samples/Example_insertOrReplace.js
@@ -1,12 +1,23 @@
+/*
+ * Copyright (c) 2018 MapR, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 const { ConnectionManager, ODate, OTime, OTimestamp } = require('node-maprdb');
 
-// Create connection with specified URL
-const connection = ConnectionManager.getConnection('localhost:5678');
-
-const storeName = '/test-db-1';
-
 const docId = 'user000';
-const store = connection.getStore(storeName);
+
 const doc = {};
 doc._id = docId;
 doc.date = new ODate('2012-10-20');
@@ -25,6 +36,14 @@ doc.phoneNumbers = [
   {areaCode: '555', number: '5555555'},
   {areaCode: '555', number: '555-5556'}
 ];
+
+// Create connection with specified connection string
+const connection = ConnectionManager.getConnection('localhost:5678');
+
+const storeName = '/test-db-1';
+
+const store = connection.getStore(storeName);
+
 store.insertOrReplace(doc)
   .then((res) => console.log(res))
   .catch((err) => console.error(err))
