+++ src/lib/DocumentStore.ts
@@ -1,6 +1,17 @@
 /*
- * Copyright (c) 2018 & onwards. MapR Tech, Inc., All rights reserved
+ * Copyright (c) 2018 MapR, Inc.
  *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 import {decode, encode, InsertOrReplaceRequestBuilder} from './Connection'
@@ -102,7 +113,7 @@ export class DocumentStore {
   public update(_id: string, mutation: any, callback?: Callback): void|Promise<any> {
     return this.grpcUpdate(_id, mutation, callback)
   }
-  public checkAndMutate(_id: string, mutation: any, condition: any, callback?: Callback): void|Promise<any> {
+  public checkAndUpdate(_id: string, mutation: any, condition: any, callback?: Callback): void|Promise<any> {
     return this.grpcUpdate(_id, mutation, condition, callback)
   }
   private grpcInsertOrReplace(
